Preliminary break down of wamr components
